Rename file according to rubocop style
Naming/FileName: The name of this source file (`yard-apptweak-template.rb`) should use snake_case. (https://rubystyle.guide#snake-case-files)
