+++ lib/yard_apptweak_template/patches/c_base_handler.rb
