+++ lib/yard_apptweak_template/stubs/autoload.rb
