+++ lib/yard_apptweak_template/templates/apis/fulldoc/text/setup.rb
