+++ lib/yard_apptweak_template/templates/default/docstring/setup.rb
