+++ lib/yard_apptweak_template/templates/default/fulldoc/html/full_list_object_types.erb
