+++ lib/yard_apptweak_template/templates/default/fulldoc/html/js/jquery-migrate.js
