+++ lib/yard_apptweak_template/templates/default/fulldoc/html/js/sketchup.js
