+++ lib/yard_apptweak_template/templates/default/layout/html/navbar.erb
