+++ lib/yard_apptweak_template/templates/default/method_details/setup.rb
