+++ lib/yard_apptweak_template/templates/default/module/html/box_info.erb
