+++ lib/yard_apptweak_template/templates/default/module/html/constant_summary.erb
