+++ lib/yard_apptweak_template/templates/inheritance/fulldoc/text/setup.rb
