+++ lib/yard_apptweak_template/templates/rubocop-changelog/fulldoc/text/setup.rb
