+++ lib/yard_apptweak_template/templates/stubs/fulldoc/text/setup.rb
