+++ lib/yard_apptweak_template/templates/versions/fulldoc/text/setup.rb
