+++ lib/yard_apptweak_template/yard/handlers/class_constants.rb
