+++ lib/yard_apptweak_template/yard/handlers/class_enum_constants.rb
