+++ lib/yard_apptweak_template/yard/handlers/global_constants.rb
